% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_occurGBIF.R
\name{clean_occurGBIF}
\alias{clean_occurGBIF}
\title{Function that read and clean GBIF occurrence file}
\usage{
clean_occurGBIF(
  ffile,
  uncertainty = 1000,
  year.min = 2000,
  cols.tokeep = c("species", "eventDate", "day", "month", "year", "X", "Y",
    "coordinateUncertaintyInMeters")
)
}
\arguments{
\item{ffile}{name of the GBIF occurrence file to read}

\item{uncertainty}{resolution in meters to filter data from, for ex., 1000 means that coordinates with uncertainty higher than 1000m are removed}

\item{year.min}{minimum year to filter data from, for ex., 2000 means that occurences before year 2000 are removed}

\item{cols.tokeep}{names of columns to keep from the GBIF file}
}
\description{
Function that read and clean GBIF occurrence file
}
