% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_griddedOccur.R
\name{get_griddedOccur}
\alias{get_griddedOccur}
\title{Function to projet GBIF occurence on a grid}
\usage{
get_griddedOccur(
  sp.occur,
  year.todo = c(2010:2022),
  grid = rr.poly,
  grid.rast = rr,
  nlim = 25,
  crs.sp.occur = 4326,
  colcoords.sp.occur = c("X", "Y"),
  yearly = F,
  save.raster = T
)
}
\arguments{
\item{sp.occur}{dataframe of geolocated occurrences}

\item{year.todo}{year of occurrences to keep}

\item{grid}{grid in polygon format}

\item{grid.rast}{grid in raster format}

\item{nlim}{minimum number of occurrences per cell to proceed the analyses}

\item{crs.sp.occur}{CRS of geolocated occurrences in sp.occur}

\item{colcoords.sp.occur}{vector of names of the XY (coordinates) columns in sp.occur}

\item{yearly}{whether to produce yearly occurrence grids, default = TRUE}

\item{save.raster}{whether to save the grids as raster (in addition to the geopackages), default = FALSE}
}
\value{
Geopackage of the grid per species with the number of occurrences per cell.
Text file is also generated to provide info on whether at least one cell contains at least "nlim" occurrences.
SIG layers are saved in data/derived-data/SIG/Occurrence/AllYearsCombined or data/derived-data/SIG/Occurrence/Yearly
Textfile is saved in data/derived-data/SIG/Occurrence/
}
\description{
Function to projet GBIF occurence on a grid
}
