% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speciesname.R
\name{get_speciesname}
\alias{get_speciesname}
\title{Function to get species name from GBIF file name}
\usage{
get_speciesname(x)
}
\arguments{
\item{x}{a GBIF file name}
}
\value{
the name of the species for the specified GBIF file name
}
\description{
Function to get species name from GBIF file name
}
